remove useMenuItemSelectable
