+++ packages/react-menu/etc/react-menu.api.md
@@ -254,9 +254,6 @@ export const useMenuItemCheckbox: (props: MenuItemCheckboxProps, ref: React.Ref<
 export const useMenuItemRadio: (props: MenuItemRadioProps, ref: React.Ref<HTMLElement>, defaultProps?: MenuItemRadioProps | undefined) => MenuItemRadioState;
 
 // @public
-export const useMenuItemSelectable: (state: MenuItemSelectableState, handleSelection?: SelectableHandler) => void;
-
-// @public
 export const useMenuItemStyles: (state: MenuItemState) => void;
 
 // @public
