+++ packages/react-menu/src/components/MenuItemRadio/useMenuItemRadio.ts
@@ -1,7 +1,6 @@
 import * as React from 'react';
 import { makeMergePropsCompat, resolveShorthandProps } from '@fluentui/react-utilities';
 import { MenuItemRadioProps, MenuItemRadioState } from './MenuItemRadio.types';
-import { useMenuItemSelectable } from '../../selectable/index';
 import { useMenuListContext } from '../../menuListContext';
 import { useMenuItem, menuItemShorthandProps } from '../MenuItem/useMenuItem';
 
@@ -29,7 +28,25 @@ export const useMenuItemRadio = (
   // set to undefined since it will be resolved again anyway
   ((baseState as unknown) as MenuItemRadioProps).checkmark = undefined;
   const state = mergeProps(baseState, defaultProps, resolveShorthandProps(props, menuItemRadioShorthandProps));
+
   const selectRadio = useMenuListContext(context => context.selectRadio);
-  useMenuItemSelectable(state, selectRadio);
+  const { onClick: onClickOriginal } = state;
+  const checked = useMenuListContext(context => {
+    const checkedItems = context.checkedValues?.[state.name] || [];
+    return checkedItems.indexOf(state.value) !== -1;
+  });
+
+  state.checked = checked;
+  state['aria-checked'] = state.checked;
+
+  state.onClick = e => {
+    if (state.disabled) {
+      return;
+    }
+
+    selectRadio?.(e, state.name, state.value, state.checked);
+    onClickOriginal?.(e);
+  };
+
   return state;
 };
